bug fix: ortho and freqstart heuristics
ortho and freqstart heuristics rely on the type-based stage which is the output
of prob_abbr.

in the context of ortho, a given token marked as an abbr by prob_abbr must also
be seen as a possible sentence ender.

in the context of freqstart, however, only the tokens that were not marked as
enders by prob_abbr are used in the colocation log-likelihood calculation.

since the freq_after_ender member of OrthoFreq is provided exclusively for the
benefit of freqstart, we only increment it when a true sentence ender (as
prob_abbr sees it) is encountered.