fix the still-unfixed ortho bug
there are in fact not two but three different positional types in a
type-classified token stream: sentence internal, sentence first, and
post-abbreviations. the first and third need to be distinguished because
strunk's ortho and freqstart specifically disregard the third (from section
4.1.1):

The Orthographic Heuristic...Punkt does not count tokens occurring after an
abbreviation or an ellipsis, because we have not yet determined whether they
start a new sentence or not. The algorithm also ignores tokens that occur after
possible initials and numbers...